:truck: Moving code into component
